+++ src/main/webapp/javascript/Menu.jsx
@@ -1,14 +1,9 @@
 import React, { useState, useEffect } from "react";
-import { BrowserRouter, Routes, Route, Outlet, Link } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 import "../css/style.css";
 
-export default function Menu(){
-    const [isActive, setIsActive] = useState(true);
-
-    const toggleActive = () => {
-        setIsActive((current) => !current);
-    }
+export default function Menu({ isActive }){
 
     if(isActive){
         return (
@@ -19,14 +14,14 @@ export default function Menu(){
                         <img src="images/animated-candle.gif"></img>
                     </div>
                     <div className="btn-container">
-                        <Link to="/Maze" >
-                            <button className="menu-btn" onClick={toggleActive}>Begin Quest</button>
+                        <Link to="/Game" >
+                            <button className="menu-btn" >Begin Quest</button>
                         </Link>
-                            <button className="menu-btn" onClick={toggleActive}>What is a regular expression?</button>
+                            <button className="menu-btn" >What is a regular expression?</button>
                         <Link to="/About">
-                            <button className="menu-btn" onClick={toggleActive}>About the Game</button>
+                            <button className="menu-btn" >About the Game</button>
                         </Link>
-                        <button className="menu-btn" onClick={toggleActive}>Quit</button>
+                        <button className="menu-btn" >Quit</button>
                     </div>
                     <div>
                         <img src="images/animated-candle.gif"></img>
